fixed and moved class to file
