+++ homeworks/vladyslav.kovalov_kovalov/4-modules/index.html
@@ -43,6 +43,6 @@
                 <button type="button" class="card__button" data-button>Weeks Number In Month</button>
             </section>
         </div>
-        <script type="module" defer src="app.js"></script>
+        <script type="module" defer src="index.js"></script>
     </body>
 </html>
